chore: Move content components
