+++ src/components/content/Markdown.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment happy-dom
-import Markdown from '$components/Markdown.svelte';
+import Markdown from '$components/content/Markdown.svelte';
 import { render } from '@testing-library/svelte';
 import { expect, it } from 'vitest';
