+++ src/components/content/QRCode.svelte
