+++ src/components/content/TocTree.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment happy-dom
-import TocTree from '$components/TocTree.svelte';
+import TocTree from '$components/content/TocTree.svelte';
 import { render } from '@testing-library/svelte';
 import { expect, it } from 'vitest';
